feat(middleware_common::metering): Dynamic limit
Add a dynamic limit to the metering::Metering middleware.

Previously a limit was hardcoded at compile time. This commit adds an
additional InternalField to track a dynamic limit unique to each
instance. This commit adds fn metering::set_limit(&mut Instance,  u64)
to allow limits to be changed.

This allows a module compiled with metering to be cached and re-used to
instantiate Instances with different limits.

implements #993